\name{udiff-package}
\alias{udiff-package}
\alias{udiff}
\docType{package}
\title{
\packageTitle{udiff}
}
\description{
\packageDescription{udiff}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{udiff}
\packageIndices{udiff}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{udiff}

Maintainer: \packageMaintainer{udiff}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
