% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udiff.R
\name{udiff}
\alias{udiff}
\title{importFrom("stats", ".getXlevels", "is.empty.model", "model.matrix",
"model.offset", "model.response", "model.weights", "optim",
"printCoefmat", "pt", "runif")}
\usage{
udiff(formula, data, subset, weights, na.action, contrasts = NULL, offset, ...)
}
\description{
importFrom("stats", ".getXlevels", "is.empty.model", "model.matrix",
"model.offset", "model.response", "model.weights", "optim",
"printCoefmat", "pt", "runif")
}
